adicionando page view model
